Renaming to mutable_robot_state_publisher so can be run in parallel with original version.
